requirements.txt
# Please use python V 3.7 to be compatible with all packages
torch>=1.9.0,<2.6
scikit-learn>=0.24.2
pyyaml>=5.4.1
numpy>=1.21.2,<=1.26.4
requests>=2.23.0
einops>=0.8.0
autogluon.tabular

#mothernet @ git+https://github.com/microsoft/ticl.git@4cb7ac0fc04aa15256995dd85f379562ea6e994c

tqdm
gpytorch
openml
hyperopt
lightgbm

huggingface_hub  # to download models
seaborn
matplotlib<=3.9.0
pandas
autorank


wandb
mlflow
interpret
mlxp

tabulate


huggingface-hub

pytest
